requirements.txt
# dependencies
asyncio==3.4.3
beautifulsoup4==4.12.2
colorama==0.4.6
duckduckgo_search==3.8.5
md2pdf==1.0.1
openai~=0.28.0
playwright==1.36.0
python-dotenv~=1.0.0
pyyaml==6.0.1
selenium
webdriver-manager==4.0.0
flask
uvicorn
pydantic
fastapi
python-multipart
markdown
langchain==0.0.275
chromedriver-autoinstaller
